feat: 프로젝트 검색 컴포넌트 위치 변경
